update test to no produce data

+++ tests/python/test_examples.py
@@ -44,28 +44,30 @@ class TestExamples(unittest.TestCase, metaclass=MetaTest):
 
     """Examples test"""
     def test_python_readme(self):
-        fname = os.path.join(BASE_DIR, "ioh", "README.md")
-        self.assertTrue(os.path.isfile(fname))
-        with open(fname) as f:
-            data = f.read().split("```")
-            
-            with io.StringIO() as buf, redirect_stdout(buf):
-                for i, x in enumerate(data):
-                    if x.startswith("python"):
-                        block = x[6:].strip()
-                        try:
-                            exec(block, GB, LC)
-                        except Exception as e:
-                            raise RuntimeError(f"failed in cell {i}, reason:\n{e}")
-        
-        shutil.rmtree("temp")
-        shutil.rmtree("temp2")
-        shutil.rmtree("temp3")
-        shutil.rmtree("ioh_data")
-        os.remove("ioh_data.zip")
-                            
-
-
+        try:
+            fname = os.path.join(BASE_DIR, "ioh", "README.md")
+            self.assertTrue(os.path.isfile(fname))
+            with open(fname) as f:
+                data = f.read().split("```")
+                
+                with io.StringIO() as buf, redirect_stdout(buf):
+                    for i, x in enumerate(data):
+                        if x.startswith("python"):
+                            block = x[6:].strip()
+                            try:
+                                exec(block, GB, LC)
+                            except Exception as e:
+                                raise RuntimeError(f"failed in cell {i}, reason:\n{e}")
+        except:
+            raise
+        finally:
+            shutil.rmtree("temp", ignore_errors=True)
+            shutil.rmtree("temp2", ignore_errors=True)
+            shutil.rmtree("temp3", ignore_errors=True)
+            shutil.rmtree("ioh_data", ignore_errors=True)
+            if os.path.exists("ioh_data.zip"):
+                os.remove("ioh_data.zip")
+                                
 
 if __name__ == "__main__":
     unittest.main()
